Add Storefront name to env variables and edit on-install script to use it

+++ .artifakt/shopware/rootfs/etc/shopware/scripts/on-init/config-fixer.sh
@@ -1,5 +1,8 @@
 #!/usr/bin/env bash
 
-rm /var/www/html/config/services/defaults_test.xml
+if [ -f /var/www/html/config/services/defaults_test.xml ]; then
+    rm /var/www/html/config/services/defaults_test.xml
+fi
+
 cp /tmp/rootfs/etc/shopware/configs/services/jwt.xml /var/www/html/config/services/
 cp /tmp/rootfs/etc/shopware/configs/services/services.xml /var/www/html/config/
